Add standard `parseStatus` helper
